chore(importForm): Commenter ImportForm
[30min][Done]

+++ PlotThoseLines/PlotThoseLines/ImportForm.cs
@@ -17,14 +17,18 @@ namespace PlotThoseLines
 {
     public partial class ImportForm : Form
     {
-        private string _filename;
-        private string _savefilename;
-        public event EventHandler SeriesImported;
+        private string _filename; //Le nom du fichier à importer
+        private string _savefilename; //Le nom du fichier à restaurer
+        public event EventHandler SeriesImported; //Evenement pour indiquer que les séries ont été importée
         public ImportForm()
         {
             InitializeComponent();
         }
 
+        /// <summary>
+        /// Import an excel file
+        /// </summary>
+        /// <returns>true if success, else false</returns>
         private bool ImportFile()
         {
             try
@@ -40,14 +44,19 @@ namespace PlotThoseLines
                     List<Serie> importedSeries = new List<Serie>();
 
                     //DataRow et DataColumn ne supportes pas le .ForEach
-                    // On all tables' rows
+                    //Première ligne du fichier: années des séries
                     List<Double> XaxisValue = table.Rows[0].ItemArray.ToList().Skip(1).Select(x => double.Parse(x.ToString())).ToList();
+
+                    //Pour chaque ligne
                     foreach (DataRow row in table.Rows)
                     {
+                        //Nom de la série
                         Serie s = new Serie(row[0].ToString(), XaxisValue, new List<double>());
-                        // On all tables' columns
+
+                        //Pour chaque colonne
                         foreach (DataColumn col in table.Columns)
                         {
+                            //Si ce n'est pas le nom de la série
                             if (col.Ordinal != 0)
                             {
                                 string cell = row[col].ToString();
@@ -68,9 +77,13 @@ namespace PlotThoseLines
                         importedSeries.Add(s);
                     }
 
+                    //On enlève les 2 premières séries (en-tête et vide) et les deux dernières (vide et copyright)
                     Program.series.AddRange(importedSeries.Skip(2).SkipLast(2).ToList());
+
+                    //Si deux séries ont le même nom et nombre d'élèment, on garde la dernière série ajoutée
                     Program.series = Program.series.GroupBy(s => s.Name + s.YaxisValue.Count).Select(s => s.Last()).ToList();
 
+                    //On sauvegarde les données dans la DB
                     SaveFile.Save();
 
                     reader.Close();
@@ -92,6 +105,11 @@ namespace PlotThoseLines
             }
         }
 
+        /// <summary>
+        /// Open the FileDialog to select the file to import
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void buttonImportFile_Click(object sender, EventArgs e)
         {
             OpenFileDialog openFileDialog = new OpenFileDialog();
@@ -109,20 +127,37 @@ namespace PlotThoseLines
             this.labeImportFile.Text = String.IsNullOrEmpty(this._filename) ? "Choisir un fichier" : this._filename;
         }
 
+        /// <summary>
+        /// Start the file import
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void buttonImport_Click(object sender, EventArgs e)
         {
-            bool success = ImportFile();
-            if (success)
+            if (!string.IsNullOrEmpty(_filename))
             {
-                SeriesImported.Invoke(this, EventArgs.Empty);
-                this.Close();
+                bool success = ImportFile();
+                if (success)
+                {
+                    SeriesImported.Invoke(this, EventArgs.Empty);
+                    this.Close();
+                }
+                else
+                {
+                    this.labelError.Text = "Erreur lors de l'import du fichier: mauvais format de fichier";
+                }
             }
             else
             {
-                this.labelError.Text = "Erreur lors de l'import du fichier: mauvais format de fichier";
+                this.labelError.Text = "Merci de choisir un fichier à importer";
             }
         }
 
+        /// <summary>
+        /// Open the FileDialog to select the file to restore
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void buttonRestoreFile_Click(object sender, EventArgs e)
         {
             OpenFileDialog openFileDialog = new OpenFileDialog();
@@ -140,32 +175,49 @@ namespace PlotThoseLines
             this.labelRestoreFile.Text = String.IsNullOrEmpty(this._savefilename) ? "Choisir un fichier" : this._savefilename;
         }
 
+        /// <summary>
+        /// Restore the selected file
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void buttonRestore_Click(object sender, EventArgs e)
         {
-            LoadingModal loader = new LoadingModal();
-            loader.Show();
+            if (!string.IsNullOrEmpty(_savefilename))
+            {
+                //Spinner de chargement
+                LoadingModal loader = new LoadingModal();
+                loader.Show();
 
-            Program.series.Clear();
-            string connectionString = "Data Source=ptl.db;Version=3;";
-            SQLiteConnection connection = new SQLiteConnection(connectionString);
-            connection.Open();
+                //Supprimemr les séries existantes de la mémoire et de la DB
+                Program.series.Clear();
+                string connectionString = "Data Source=ptl.db;Version=3;";
+                SQLiteConnection connection = new SQLiteConnection(connectionString);
+                connection.Open();
 
-            new SQLiteCommand("DELETE FROM points", connection).ExecuteNonQuery();
-            new SQLiteCommand("DELETE FROM series", connection).ExecuteNonQuery();
+                new SQLiteCommand("DELETE FROM points", connection).ExecuteNonQuery();
+                new SQLiteCommand("DELETE FROM series", connection).ExecuteNonQuery();
 
-            string[] importSql = File.ReadAllLines(this._savefilename);
+                //Executer chaque commande du dump
+                string[] importSql = File.ReadAllLines(this._savefilename);
 
-            importSql.ToList().ForEach(x => Trace.WriteLine(x));
-            importSql.ToList().ForEach(x => new SQLiteCommand(x, connection).ExecuteNonQuery());
-            SaveFile.Load();
-            connection.Close();
+                importSql.ToList().ForEach(x => Trace.WriteLine(x));
+                importSql.ToList().ForEach(x => new SQLiteCommand(x, connection).ExecuteNonQuery());
+                SaveFile.Load();
+                connection.Close();
 
-            loader.Close();
+                loader.Close();
 
-            SeriesImported.Invoke(this, EventArgs.Empty);
-            this.Close();
+                SeriesImported.Invoke(this, EventArgs.Empty);
+                this.Close();
+            }
+            this.labelError.Text = "Merci de choisir un fichier à restaurer";
         }
 
+        /// <summary>
+        /// Close this forms
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void buttonClose_Click(object sender, EventArgs e)
         {
             this.Close();
